+++ react-onsenui/CHANGELOG.md
@@ -2,6 +2,13 @@
 CHANGELOG
 ====
 
+dev
+---
+
+### Bug Fixes
+
+ * Input, SearchInput: Fix value prop being ignored after input is touched. ([#2996](https://github.com/OnsenUI/OnsenUI/issues/2996)).
+
 1.12.0
 ---